Rename files and get rid of unneeded files
