--- src/tests/load/artillery/dailyData.ts
@@ -1,9 +1,0 @@
-import { Page } from '@playwright/test';
-
-import { testDailyData } from '@/tests/load/commands/testDailyData';
-
-async function artilleryScript(page: Page) {
-  await testDailyData(page);
-}
-
-export { artilleryScript };
