+++ src/tests/load/artillery/load.ts
@@ -1,9 +1,9 @@
 import { Page } from '@playwright/test';
 
-import { testDailyData } from '@/tests/load/commands/testDailyData';
+import { testDailyData } from '@/tests/load/commands/test-daily-data';
+import { testSleep } from '@/tests/load/commands/test-sleep';
 import { testDataVisualization } from '@/tests/load/commands/testDataVisualization';
 import { testLogin } from '@/tests/load/commands/testLogin';
-import { testSleep } from '@/tests/load/commands/testSleep';
 
 async function artilleryScript(page: Page) {
   // comment out tests you dont want running to test your own test
