+++ src/tests/load/commands/test-daily-data.ts
